+++ Nos/Views/PagedNoteListView.swift
@@ -25,6 +25,8 @@ struct PagedNoteListView<Header: View, EmptyPlaceholder: View>: UIViewRepresenta
     /// by the `databaseFilter`. 
     let relayFilter: Filter
     
+    let relay: Relay? 
+    
     let context: NSManagedObjectContext
 
     /// The tab in which this PagedNoteListView appears.
@@ -55,6 +57,7 @@ struct PagedNoteListView<Header: View, EmptyPlaceholder: View>: UIViewRepresenta
         let dataSource = context.coordinator.dataSource(
             databaseFilter: databaseFilter, 
             relayFilter: relayFilter,
+            relay: relay,
             collectionView: collectionView, 
             context: self.context,
             header: header,
@@ -81,8 +84,17 @@ struct PagedNoteListView<Header: View, EmptyPlaceholder: View>: UIViewRepresenta
         return collectionView
     }
     
-    func updateUIView(_ collectionView: UICollectionView, context: Context) {}
-
+    func updateUIView(_ collectionView: UICollectionView, context: Context) {
+        if let dataSource = collectionView.dataSource as? PagedNoteDataSource<Header, EmptyPlaceholder> {
+            if relayFilter != dataSource.relayFilter || relay != dataSource.relay {
+                dataSource.subscribeToEvents(matching: relayFilter, from: relay)
+            }
+            if databaseFilter != dataSource.databaseFilter {
+                dataSource.updateFetchRequest(databaseFilter)
+            }
+        }
+    }
+    
     static func dismantleUIView(_ uiView: UICollectionView, coordinator: Coordinator<Header, EmptyPlaceholder>) {
         tearDownObservers(coordinator: coordinator)
     }
@@ -164,6 +176,7 @@ struct PagedNoteListView<Header: View, EmptyPlaceholder: View>: UIViewRepresenta
         func dataSource(
             databaseFilter: NSFetchRequest<Event>, 
             relayFilter: Filter,
+            relay: Relay?,
             collectionView: UICollectionView,
             context: NSManagedObjectContext,
             @ViewBuilder header: @escaping () -> CoordinatorHeader,
@@ -179,6 +192,7 @@ struct PagedNoteListView<Header: View, EmptyPlaceholder: View>: UIViewRepresenta
             let dataSource = PagedNoteDataSource(
                 databaseFilter: databaseFilter, 
                 relayFilter: relayFilter,
+                relay: relay,
                 collectionView: collectionView, 
                 context: context,
                 header: header,
@@ -192,7 +206,6 @@ struct PagedNoteListView<Header: View, EmptyPlaceholder: View>: UIViewRepresenta
         @objc func refreshData(_ sender: Any) {
             if let onRefresh {
                 dataSource?.updateFetchRequest(onRefresh())
-                collectionView?.reloadData()
             }
             
             if let refreshControl = sender as? UIRefreshControl {
@@ -217,7 +230,8 @@ extension Notification.Name {
     
     return PagedNoteListView(
         databaseFilter: previewData.alice.allPostsRequest(onlyRootPosts: false),
-        relayFilter: Filter(keepSubscriptionOpen: true),
+        relayFilter: Filter(), 
+        relay: nil,
         context: previewData.previewContext,
         tab: .home,
         header: {